requirements.txt
tweepy
python-dotenv
langchain
ctransformers
langchain-community